\name{diffGOStats}
\alias{diffGOStats}
\title{calculate difference between noise and clean differences for 100 gene samples}
\usage{
  diffGOStats(inList)
}
\arguments{
  \item{inList}{the list of results to work with}
}
\value{
  data.frame
}
\description{
  calculate difference between noise and clean differences
  for 100 gene samples
}

