\name{fishersMethod}
\alias{fishersMethod}
\title{fishers method for combining p-values}
\usage{
  fishersMethod(x)
}
\arguments{
  \item{x}{set of p-values to combine}
}
\value{
  combined p-values
}
\description{
  fishers method for combining p-values
}
\details{
  function from
  \url{http://mikelove.wordpress.com/2012/03/12/combining-p-values-fishers-method-sum-of-p-values-binomial/}
}

