\name{gseaSizeRange}
\alias{gseaSizeRange}
\title{generate entries within a distribution}
\usage{
  gseaSizeRange(inLimits, nSample)
}
\arguments{
  \item{inLimits}{the range to use}

  \item{nSample}{how many entries within the range}
}
\value{
  set of sizes
}
\description{
  generate entries within a distribution
}

