\docType{data}
\name{lung.RData}
\alias{lung.RData}
\title{lung.RData}
\source{
  ftp://ftp.ncbi.nlm.nih.gov/geo/series/GSE18nnn/GSE18842/matrix/GSE18842_series_matrix.txt.gz
}
\description{
  lung.RData
}
\details{
  Downloaded, created an \code{ExpressionSet} object using
  \code{getGEO} on Dec 6, 2013.
}
\keyword{datasets}

