\name{mmGeneSetTest}
\alias{mmGeneSetTest}
\title{multiple-multiple geneSetTests}
\usage{
  mmGeneSetTest(index, statistics, ranks.only = FALSE,
    do.parallel = TRUE, ...)
}
\arguments{
  \item{index}{list of indices generated by
  \code{symbols2indices}}

  \item{statistics}{the statistics to use}

  \item{ranks.only}{whether to rank only, default is FALSE}

  \item{do.parallel}{use multiple cores to do everything}

  \item{...}{other parameters for \code{geneSetTest}}
}
\value{
  data.frame of p-values for each alternative test
}
\description{
  The default \code{geneSetTest} in \code{limma} will only
  do a single index and a single test. This function allows
  one to pass in a set of indices, and do all three tests
  (up, down, mixed).
}
\details{
  Note that the function will try to do all three of up,
  down and mixed, but if there is no sign (i.e. F-statistic
  instead of t-statistic) the function will only do a test
  for "mixed"
}

