\docType{data}
\name{muscle_data.RData}
\alias{muscle_data.RData}
\title{muscle_data.RData}
\source{
  http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE4411
}
\description{
  muscle_data.RData
}
\details{
  RMA normalized intensities downloaded from GEO using
  \code{getGEO}, on January 13, 2014; data was trimmed down
  to only the denervated and innervated controls
}
\keyword{datasets}

