\docType{data}
\name{muscleGeneSetTest.RData}
\alias{muscleGeneSetTest.RData}
\title{muscleGeneSetTest.RData}
\source{
  running \code{mmGeneSetTest} on muscle denervated over
  innervated
}
\description{
  muscleGeneSetTest.RData
}
\details{
  did one contrast, denervated vs innervated using the
  \code{mmGeneSetTest} function
}
\keyword{datasets}

