\docType{data}
\name{muscleRomer.RData}
\alias{muscleRomer.RData}
\title{muscleRomer.RData}
\source{
  running \code{romer} on muscle denervated over innervated
}
\description{
  muscleRomer.RData
}
\details{
  did one contrast, denervated vs innervated using the
  \code{multicontrastRomer} function
}
\keyword{datasets}

