\docType{data}
\name{noiseSamples.RData}
\alias{noiseSamples.RData}
\title{noiseSamples.RData}
\source{
  hypothetical vignette from \package{ccPaperRev}
}
\description{
  noiseSamples.RData
}
\details{
  For two 1000 gene samples, different amounts of noise (0
  - 1000) genes (genes not annotated to 100 GO terms
  defining the samples) were added, with different
  fractions of overlap between the noise gene lists (0 -
  1).
}
\keyword{datasets}

