\name{possibleNoise}
\alias{possibleNoise}
\title{possible noise genes to sample from}
\usage{
  possibleNoise(go2gene, goSample)
}
\arguments{
  \item{go2gene}{list of GO term to gene mapping}

  \item{goSample}{character list of GO terms already
  sampled}
}
\value{
  noise genes to sample from
}
\description{
  possible noise genes to sample from
}

