\name{pvalueDiffSig}
\alias{pvalueDiffSig}
\title{calculate difference and significance}
\usage{
  pvalueDiffSig(pvalueData, pCutoff = 0.05, log = TRUE)
}
\arguments{
  \item{pvalueData}{the \code{data.frame} of values}

  \item{pCutoff}{the p-value cutoff to use for
  significance}

  \item{log}{the values are logged, and p-value should be
  too}
}
\value{
  \code{data.frame}
}
\description{
  calculate difference and significance
}

