\docType{data}
\name{skinGeneSetTest.RData}
\alias{skinGeneSetTest.RData}
\title{skinGeneSetTest.RData}
\source{
  running \code{geneSetTest} on skin t7 vs t0 and t14 vs t0
  comparisons
}
\description{
  skinGeneSetTest.RData
}
\details{
  did two sets of contrasts, t7 vs t0 and t14 vs t0, and
  up, down, and mixed alternative tests using
  \code{mmGeneSetTest}
}
\keyword{datasets}

