\docType{data}
\name{skinRomer.RData}
\alias{skinRomer.RData}
\title{skinRomer.RData}
\source{
  running \code{romer} on skin t7 vs t0 and t14 vs t0
  comparisons
}
\description{
  skinRomer.RData
}
\details{
  did two sets of contrasts, t7 vs t0 and t14 vs t0 using
  the \code{multicontrastRomer} function
}
\keyword{datasets}

