\docType{data}
\name{uc_crohns_rawData.RData}
\alias{uc_crohns_rawData.RData}
\title{uc_crohns_rawData.RData}
\source{
  http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE36807
}
\description{
  uc_crohns_rawData.RData
}
\details{
  Used getGEO to get data on January 13, 2014
}
\keyword{datasets}

